Include platform specific `AssetWatcher`
Depending on the target platform, include the appropriate
`AssetWatcher` implementation, defaulting to a simple 5 second
timer.
